ASIC bus - read only.  First attempt.
Updated 2 minutes ago 


Arduino code to monitor the bus and identify response telegraphs that contain the “present value” data for selected analogue inputs. The target data is extracted and displays it on the serial monitor. It does NOT make the request for the data. Instead, it intercepts responses intended for the HMI panel. This was good enough for a first attempt.